+++ src/core/receiver/gnss_block_factory.cc
@@ -167,8 +167,17 @@
 #include "fmcomms2_signal_source.h"
 #endif
 
-#if AD9361_DRIVER
-#include "ad9361_fpga_signal_source.h"
+#if ENABLE_FPGA and AD9361_DRIVER
+#include "adrv9361_z7035_signal_source_fpga.h"
+#include "fmcomms5_signal_source_fpga.h"
+#endif
+
+#if MAX2771_DRIVER
+#include "max2771_evkit_signal_source_fpga.h"
+#endif
+
+#if DMA_PROXY_DRIVER
+#include "dma_signal_source_fpga.h"
 #endif
 
 #if LIMESDR_DRIVER
@@ -812,12 +821,34 @@ std::unique_ptr<GNSSBlockInterface> GNSSBlockFactory::GetBlock(
                 }
 #endif
 
-#if AD9361_DRIVER
-            // The AD9361_DRIVER Driver must be instantiated last. In this way, when using the FPGA, and when using the GNSS receiver
-            // in post-processing mode, the receiver is configured and ready when the DMA starts sending samples to the receiver.
-            else if (implementation == "Ad9361_Fpga_Signal_Source")
+#if ENABLE_FPGA and AD9361_DRIVER
+            else if (implementation == "ADRV9361_Z7035_Signal_Source_FPGA")
                 {
-                    std::unique_ptr<GNSSBlockInterface> block_ = std::make_unique<Ad9361FpgaSignalSource>(configuration, role, in_streams,
+                    std::unique_ptr<GNSSBlockInterface> block_ = std::make_unique<Adrv9361z7035SignalSourceFPGA>(configuration, role, in_streams,
+                        out_streams, queue);
+                    block = std::move(block_);
+                }
+            else if (implementation == "FMCOMMS5_Signal_Source_FPGA")
+                {
+                    std::unique_ptr<GNSSBlockInterface> block_ = std::make_unique<Fmcomms5SignalSourceFPGA>(configuration, role, in_streams,
+                        out_streams, queue);
+                    block = std::move(block_);
+                }
+#endif
+
+#if ENABLE_FPGA and MAX2771_DRIVER
+            else if (implementation == "MAX2771_EVKIT_Signal_Source_FPGA")
+                {
+                    std::unique_ptr<GNSSBlockInterface> block_ = std::make_unique<MAX2771EVKITSignalSourceFPGA>(configuration, role, in_streams,
+                        out_streams, queue);
+                    block = std::move(block_);
+                }
+#endif
+
+#if ENABLE_FPGA and DMA_PROXY_DRIVER
+            else if (implementation == "DMA_Signal_Source_FPGA")
+                {
+                    std::unique_ptr<GNSSBlockInterface> block_ = std::make_unique<DMASignalSourceFPGA>(configuration, role, in_streams,
                         out_streams, queue);
                     block = std::move(block_);
                 }
@@ -1054,31 +1085,31 @@ std::unique_ptr<GNSSBlockInterface> GNSSBlockFactory::GetBlock(
                 }
 #endif
 #if ENABLE_FPGA
-            else if (implementation == "GPS_L1_CA_PCPS_Acquisition_Fpga")
+            else if (implementation == "GPS_L1_CA_PCPS_Acquisition_FPGA")
                 {
                     std::unique_ptr<GNSSBlockInterface> block_ = std::make_unique<GpsL1CaPcpsAcquisitionFpga>(configuration, role, in_streams,
                         out_streams);
                     block = std::move(block_);
                 }
-            else if (implementation == "Galileo_E1_PCPS_Ambiguous_Acquisition_Fpga")
+            else if (implementation == "Galileo_E1_PCPS_Ambiguous_Acquisition_FPGA")
                 {
                     std::unique_ptr<GNSSBlockInterface> block_ = std::make_unique<GalileoE1PcpsAmbiguousAcquisitionFpga>(configuration, role, in_streams,
                         out_streams);
                     block = std::move(block_);
                 }
-            else if (implementation == "GPS_L2_M_PCPS_Acquisition_Fpga")
+            else if (implementation == "GPS_L2_M_PCPS_Acquisition_FPGA")
                 {
                     std::unique_ptr<GNSSBlockInterface> block_ = std::make_unique<GpsL2MPcpsAcquisitionFpga>(configuration, role, in_streams,
                         out_streams);
                     block = std::move(block_);
                 }
-            else if (implementation == "GPS_L5i_PCPS_Acquisition_Fpga")
+            else if (implementation == "GPS_L5i_PCPS_Acquisition_FPGA")
                 {
                     std::unique_ptr<AcquisitionInterface> block_ = std::make_unique<GpsL5iPcpsAcquisitionFpga>(configuration, role, in_streams,
                         out_streams);
                     block = std::move(block_);
                 }
-            else if (implementation == "Galileo_E5a_Pcps_Acquisition_Fpga")
+            else if (implementation == "Galileo_E5a_Pcps_Acquisition_FPGA")
                 {
                     std::unique_ptr<GNSSBlockInterface> block_ = std::make_unique<GalileoE5aPcpsAcquisitionFpga>(configuration, role, in_streams,
                         out_streams);
@@ -1204,31 +1235,31 @@ std::unique_ptr<GNSSBlockInterface> GNSSBlockFactory::GetBlock(
                 }
 #endif
 #if ENABLE_FPGA
-            else if (implementation == "GPS_L1_CA_DLL_PLL_Tracking_Fpga")
+            else if (implementation == "GPS_L1_CA_DLL_PLL_Tracking_FPGA")
                 {
                     std::unique_ptr<TrackingInterface> block_ = std::make_unique<GpsL1CaDllPllTrackingFpga>(configuration, role, in_streams,
                         out_streams);
                     block = std::move(block_);
                 }
-            else if (implementation == "Galileo_E1_DLL_PLL_VEML_Tracking_Fpga")
+            else if (implementation == "Galileo_E1_DLL_PLL_VEML_Tracking_FPGA")
                 {
                     std::unique_ptr<TrackingInterface> block_ = std::make_unique<GalileoE1DllPllVemlTrackingFpga>(configuration, role, in_streams,
                         out_streams);
                     block = std::move(block_);
                 }
-            else if (implementation == "GPS_L2_M_DLL_PLL_Tracking_Fpga")
+            else if (implementation == "GPS_L2_M_DLL_PLL_Tracking_FPGA")
                 {
                     std::unique_ptr<GNSSBlockInterface> block_ = std::make_unique<GpsL2MDllPllTrackingFpga>(configuration, role, in_streams,
                         out_streams);
                     block = std::move(block_);
                 }
-            else if ((implementation == "GPS_L5i_DLL_PLL_Tracking_Fpga") or (implementation == "GPS_L5_DLL_PLL_Tracking_Fpga"))
+            else if ((implementation == "GPS_L5i_DLL_PLL_Tracking_FPGA") or (implementation == "GPS_L5_DLL_PLL_Tracking_FPGA"))
                 {
                     std::unique_ptr<GNSSBlockInterface> block_ = std::make_unique<GpsL5DllPllTrackingFpga>(configuration, role, in_streams,
                         out_streams);
                     block = std::move(block_);
                 }
-            else if (implementation == "Galileo_E5a_DLL_PLL_Tracking_Fpga")
+            else if (implementation == "Galileo_E5a_DLL_PLL_Tracking_FPGA")
                 {
                     std::unique_ptr<GNSSBlockInterface> block_ = std::make_unique<GalileoE5aDllPllTrackingFpga>(configuration, role, in_streams,
                         out_streams);
@@ -1492,31 +1523,31 @@ std::unique_ptr<AcquisitionInterface> GNSSBlockFactory::GetAcqBlock(
         }
 #endif
 #if ENABLE_FPGA
-    else if (implementation == "GPS_L1_CA_PCPS_Acquisition_Fpga")
+    else if (implementation == "GPS_L1_CA_PCPS_Acquisition_FPGA")
         {
             std::unique_ptr<AcquisitionInterface> block_ = std::make_unique<GpsL1CaPcpsAcquisitionFpga>(configuration, role, in_streams,
                 out_streams);
             block = std::move(block_);
         }
-    else if (implementation == "Galileo_E1_PCPS_Ambiguous_Acquisition_Fpga")
+    else if (implementation == "Galileo_E1_PCPS_Ambiguous_Acquisition_FPGA")
         {
             std::unique_ptr<AcquisitionInterface> block_ = std::make_unique<GalileoE1PcpsAmbiguousAcquisitionFpga>(configuration, role, in_streams,
                 out_streams);
             block = std::move(block_);
         }
-    else if (implementation == "GPS_L2_M_PCPS_Acquisition_Fpga")
+    else if (implementation == "GPS_L2_M_PCPS_Acquisition_FPGA")
         {
             std::unique_ptr<AcquisitionInterface> block_ = std::make_unique<GpsL2MPcpsAcquisitionFpga>(configuration, role, in_streams,
                 out_streams);
             block = std::move(block_);
         }
-    else if (implementation == "GPS_L5i_PCPS_Acquisition_Fpga")
+    else if (implementation == "GPS_L5i_PCPS_Acquisition_FPGA")
         {
             std::unique_ptr<AcquisitionInterface> block_ = std::make_unique<GpsL5iPcpsAcquisitionFpga>(configuration, role, in_streams,
                 out_streams);
             block = std::move(block_);
         }
-    else if (implementation == "Galileo_E5a_Pcps_Acquisition_Fpga")
+    else if (implementation == "Galileo_E5a_Pcps_Acquisition_FPGA")
         {
             std::unique_ptr<AcquisitionInterface> block_ = std::make_unique<GalileoE5aPcpsAcquisitionFpga>(configuration, role, in_streams,
                 out_streams);
@@ -1660,31 +1691,31 @@ std::unique_ptr<TrackingInterface> GNSSBlockFactory::GetTrkBlock(
         }
 #endif
 #if ENABLE_FPGA
-    else if (implementation == "GPS_L1_CA_DLL_PLL_Tracking_Fpga")
+    else if (implementation == "GPS_L1_CA_DLL_PLL_Tracking_FPGA")
         {
             std::unique_ptr<TrackingInterface> block_ = std::make_unique<GpsL1CaDllPllTrackingFpga>(configuration, role, in_streams,
                 out_streams);
             block = std::move(block_);
         }
-    else if (implementation == "Galileo_E1_DLL_PLL_VEML_Tracking_Fpga")
+    else if (implementation == "Galileo_E1_DLL_PLL_VEML_Tracking_FPGA")
         {
             std::unique_ptr<TrackingInterface> block_ = std::make_unique<GalileoE1DllPllVemlTrackingFpga>(configuration, role, in_streams,
                 out_streams);
             block = std::move(block_);
         }
-    else if (implementation == "GPS_L2_M_DLL_PLL_Tracking_Fpga")
+    else if (implementation == "GPS_L2_M_DLL_PLL_Tracking_FPGA")
         {
             std::unique_ptr<TrackingInterface> block_ = std::make_unique<GpsL2MDllPllTrackingFpga>(configuration, role, in_streams,
                 out_streams);
             block = std::move(block_);
         }
-    else if ((implementation == "GPS_L5i_DLL_PLL_Tracking_Fpga") or (implementation == "GPS_L5_DLL_PLL_Tracking_Fpga"))
+    else if ((implementation == "GPS_L5i_DLL_PLL_Tracking_FPGA") or (implementation == "GPS_L5_DLL_PLL_Tracking_FPGA"))
         {
             std::unique_ptr<TrackingInterface> block_ = std::make_unique<GpsL5DllPllTrackingFpga>(configuration, role, in_streams,
                 out_streams);
             block = std::move(block_);
         }
-    else if (implementation == "Galileo_E5a_DLL_PLL_Tracking_Fpga")
+    else if (implementation == "Galileo_E5a_DLL_PLL_Tracking_FPGA")
         {
             std::unique_ptr<TrackingInterface> block_ = std::make_unique<GalileoE5aDllPllTrackingFpga>(configuration, role, in_streams,
                 out_streams);