+++ src/tests/unit-tests/signal-processing-blocks/acquisition/gps_l1_ca_pcps_acquisition_test_fpga.cc
@@ -66,7 +66,7 @@ class GpsL1CaPcpsAcquisitionTestFpga : public ::testing::Test
 {
 public:
     bool acquire_signal();
-    std::string implementation = "GPS_L1_CA_DLL_PLL_Tracking_Fpga";
+    std::string implementation = "GPS_L1_CA_DLL_PLL_Tracking_FPGA";
     std::vector<Gnss_Synchro> gnss_synchro_vec;
 
     const int32_t TEST_ACQ_SKIP_SAMPLES = 1024;
@@ -320,7 +320,7 @@ bool GpsL1CaPcpsAcquisitionTestFpga::acquire_signal()
     // instantiate the FPGA switch and set the
     // switch position to DMA.
     std::shared_ptr<Fpga_Switch> switch_fpga;
-    switch_fpga = std::make_shared<Fpga_Switch>("/dev/uio1");
+    switch_fpga = std::make_shared<Fpga_Switch>();
     switch_fpga->set_switch_position(0);  // set switch position to DMA
 
     // create the correspondign acquisition block according to the desired tracking signal
@@ -396,7 +396,7 @@ bool GpsL1CaPcpsAcquisitionTestFpga::acquire_signal()
 void GpsL1CaPcpsAcquisitionTestFpga::init()
 {
     config->set_property("GNSS-SDR.internal_fs_sps", "4000000");
-    config->set_property("Acquisition.implementation", "GPS_L1_CA_PCPS_Acquisition_Fpga");
+    config->set_property("Acquisition.implementation", "GPS_L1_CA_PCPS_Acquisition_FPGA");
     config->set_property("Acquisition.threshold", "0.00001");
     config->set_property("Acquisition.doppler_max", std::to_string(doppler_max));
     config->set_property("Acquisition.doppler_step", std::to_string(doppler_step));